Add new 'label' wave modifier
This is a wavedrompy-specific extension that does not exist in
wavedrom.

Add a label on a wave. It is specified along with the name, wave,
etc. of the wave signal. It contains one entry per tick, where:

  - the '.' character does not add a label

  - any other character adds this character as label

  - A string label can be encapsulated as "{abc}" for a tock

The first two are identical to node, except a label can be added
multiple times.

Furthermore the position of the label inside the tock can be modified
by adding a "(..)" modifier immediately after the label. It moves the
label relative in the tock, so that the specified offset must be
between 0 and 1.