Skip upstream comparison on older Python versions
There is a problem with xmldiff in older Python versions (<3.6) that
lead to a situation where circular UpdateMove actions are incorrectly
generated for nodes with identical children. While those are
technically okay, they are a nightmare to handle.

It should be okay to check if the tests render properly on older
versions, and only compare to upstream for newer versions.